IMPORTANT: Declare the response class
Instead of returning a `RedirectResponse` instance directly, we should
specify to FastAPI which response class to use. This allows the
framework to do the Response mapping magic it does. If you return an
explicit Response instance, it will not copy the cookies/headers from
the Response dependency that we're setting up in our dependency.